Refactoring numeric tests into a distinct folder
